Percentage of Coping Methods Used for Anxiety and Depression by Country, 2020

Percentage of Coping Methods Used for Anxiety and Depression by Country, 2020

Usage

percent_copingmethods

Format

'percent_copingmethods'

A data frame with 113 rows and 11 columns

Entity

Name of Country

Code

County Code

Year

Year of Data Collection

Religious spiritual activities

Percentage of respondents in each demographic group (country, age, gender) who engaged in religious or spiritual activities

Improved healthy lifestyle behaviors

Percentage of respondents in each demographic group (country, age, gender) who improved healthy lifestyle behaviors

Made a change to work situation

Percentage of respondents in each demographic group (country, age, gender) who made a change to work situation

Made a change to personal relationships

Percentage of respondents in each demographic group (country, age, gender) who made a change to personal relationships

Talked to friends or family

Percentage of respondents in each demographic group (country, age, gender) who spoke to friends or family

Took prescribed medication

Percentage of respondents in each demographic group (country, age, gender) who took prescribed medication

Spent time in nature

Percentage of respondents in each demographic group (country, age, gender) who spent time in nature

Talked to mental health professional

Percentage of respondents in each demographic group (country, age, gender) who spoke with a mental health professional

@source https://ourworldindata.org/grapher/dealing-with-anxiety-depression-comparison?tab=table